requirements.txt
langchain
langchain-community
langchain-experimental
fastapi==0.128.0
python-dotenv==1.2.1
streamlit==1.52.2
uvicorn==0.40.0
pydantic==2.12.5
httpx==0.28.1
requests==2.32.5
langchain-google-community[places]
langchain-tavily
langchain-google-genai
google-generativeai
langgraph
